feat: [IA-310] Onboarding PayPal files refactoring (#3449)
* refactoring

* refactoring
