+++ ts/features/wallet/onboarding/paypal/components/PspRadioItem.tsx
@@ -2,15 +2,15 @@
 import React from "react";
 import { Dimensions, Image, StyleSheet } from "react-native";
 import { View } from "native-base";
-import I18n from "../../../../i18n";
-import { useIOBottomSheetRaw } from "../../../../utils/bottomSheet";
-import { useImageResize } from "../../onboarding/bancomat/screens/hooks/useImageResize";
-import { H4 } from "../../../../components/core/typography/H4";
-import TouchableDefaultOpacity from "../../../../components/TouchableDefaultOpacity";
-import IconFont from "../../../../components/ui/IconFont";
-import { IOColors } from "../../../../components/core/variables/IOColors";
+import I18n from "../../../../../i18n";
+import { useIOBottomSheetRaw } from "../../../../../utils/bottomSheet";
+import { useImageResize } from "../../bancomat/screens/hooks/useImageResize";
+import { H4 } from "../../../../../components/core/typography/H4";
+import TouchableDefaultOpacity from "../../../../../components/TouchableDefaultOpacity";
+import IconFont from "../../../../../components/ui/IconFont";
+import { IOColors } from "../../../../../components/core/variables/IOColors";
 import { PayPalPsp } from "../screen/PayPalPspSelectionScreen";
-import { TestID } from "../../../../types/WithTestID";
+import { TestID } from "../../../../../types/WithTestID";
 import { PspInfoBottomSheetContent } from "./PspInfoBottomSheet";
 
 const PSP_LOGO_MAX_WIDTH = Dimensions.get("window").width;
