+++ ts/features/wallet/onboarding/paypal/screen/__tests__/PspRadioItem.test.tsx
@@ -3,7 +3,7 @@ import { fireEvent, render } from "@testing-library/react-native";
 import React from "react";
 import { PspRadioItem } from "../../components/PspRadioItem";
 import { PayPalPsp } from "../PayPalPspSelectionScreen";
-import { privacyUrl } from "../../../../../config";
+import { privacyUrl } from "../../../../../../config";
 
 const mockPresent = jest.fn();
 jest.mock("@gorhom/bottom-sheet", () => ({
